+++ backend/canisters/user/impl/src/lifecycle/post_upgrade.rs
@@ -94,6 +94,11 @@ fn post_upgrade(args: Args) {
         "Migrated streak insurance payments and claims to stable memory"
     );
 
+    // Move the contacts into stable memory
+    // TODO: Remove this after next release
+    let contacts_migrated = data.contacts.migrate_to_stable_memory();
+    info!(contacts_migrated, "Migrated contacts to stable memory");
+
     let env = Box::new(CanisterEnv::new(data.rng_seed));
     init_state(env, data, args.wasm_version);